backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0.25
requests==2.31.0
beautifulsoup4==4.12.0
python-dotenv==1.0.0
APScheduler==3.10.4
Flask-CORS==4.0.0
python-dateutil==2.8.2
pytest==7.4.0

pypdf==4.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
